Moved svg icons to their own files
